// src/components/_SourceTable/interface.ts
export type ITableDataSource = any[];

// @todo total footer
export interface IInputTypeColDef extends IColDef {
  inputType: InputType;
}

export interface IExtraActionEvent extends Event {
  tableDataSource: any[];
  tableFormData: any;
}

export type IExtraActionButtonOnClick = (event: IExtraActionEvent) => void;

export interface IExtraActionProps {
  onClick?: IExtraActionButtonOnClick;
}

export interface SourceTableGetContextMenuItemsParams extends TableGetContextMenuItemsParams {
  searchFormData: {};
  tableDataSource: any[];
  tableFormData: {};
  pagination: {};
}

export interface ITableFormEvent {
  formData: {};
  changed: {};
  oldFormData?: {};
  searchFormData: {};
  tableFormData?: {};
  tableDataSource: any[];
}

export type IStateRowActionButtonOnClick = (
  event: TableRowEventData & {
    state: SourceTableState;
  }
) => void;

export interface IStateRowActionProps {
  onClick?: IStateRowActionButtonOnClick;
  loading?: boolean;
}

export type IStateTableRowActions = Array<React.ReactElement<IStateRowActionProps>>;

export type IStateRowActions = Array<React.ReactElement<IStateTableRowActions>>;

export type IRowActions = Array<React.ReactElement<IRowActionProps>>;

export type ISourceTableBaseRowActions = IRowActions | ((event: TableRowEventData) => IRowActions);

export type ISourceTableBasePagination = boolean | ITablePagination;

export interface SourceTableBaseProps {
  /* 是否在底部也显示操作栏 */
  dobuleAction?: boolean;
  pagination?: ISourceTableBasePagination;
  paginationProps?: ITablePaginationProps;
  onPaginationChange?: (page: number, pageSize?: number) => void;
  onPaginationShowSizeChange?: (current: number, size: number) => void;
  resetLoading?: boolean;
  rowKey: string;
  onReset?: (
    event: {
      searchFormData: object;
      tableDataSource: any[];
      tableFormData: object;
    }
  ) => void;
  resetButtonProps?: ButtonProps;
  style?: CSSProperties;
  tableFormControls?: IFormControl[];
  onTableFormChange?: (event: ITableFormEvent) => void;
  searchText?: string;
  searchButtonProps?: ButtonProps;
  onSearch?: (
    event: {
      searchFormData: object;
      tableDataSource: any[];
      tableFormData: object;
    }
  ) => void;
  searchFormData?: object;
  dataSource?: any[];
  tableFormData?: object;
  saveText?: string;
  saveDisabled?: boolean;
  saveButtonProps?: ButtonProps;
  onSave?: (
    event: {
      searchFormData: object;
      tableDataSource: any[];
      tableFormData: {};
    }
  ) => void;
  createFormControls?: IFormControl[];
  createFormData?: object;
  onCreate?: (modalFormData?: object) => void | boolean | Promise<void | boolean | object>;
  createFormProps?: Form2StateProps;
  onSearchFormChange?: (event: ITableFormEvent) => void;
  searchFormProps?: Form2StateProps;
  searchFormControls?: IFormControl[];
  searchable?: boolean;
  editable?: boolean;
  tableColumnDefs?: IColumnDef[];
  createable?: boolean;
  createButton?: React.ReactNode;
  createModalContent?: React.ReactNode;
  tableProps?: Omit<TableProps, 'rowKey'>;
  removeable?: boolean;
  createText?: string;
  createModalProps?: ModalProps;
  onSwitchCreateModal?: IModalButtonOnSwitchModal;
  resetable?: false;
  tableFormProps?: Form2StateProps;
  insertable?: boolean;
  actionColDef?: IColumnDef;
  loading?: boolean;
  onCreateFormChange?: (
    values: object,
    tableData: any[],
    tableFormData: object,
    changed: {}
  ) => void;
  saveLoading?: boolean;
  resetText?: string;
  onInsert?: (
    event: {
      rowId: string;
      rowData: any;
      rowIndex: number;
      originEvent: Event;
    }
  ) => void;
  onRemove?: (
    event: {
      rowId: string;
      rowData: any;
      rowIndex: number;
      originEvent: Event;
    }
  ) => void;
  rowActions?: ISourceTableBaseRowActions;
  extraActions?: Array<React.ReactElement<IExtraActionProps>>;
  totalFooter?: boolean;
  totalColDef?: IColumnDef;
  createButtonProps?: ButtonProps;
}

export interface TableRowEventData {
  rowId: string;
  rowData: any;
  rowIndex: number;
  originEvent?: Event;
  params: ActionCellRendererClassParams;
}